project_amazon/requirements.txt
django
djangorestframework
django-environ
requests
pillow
psycopg2-binary

